+++ CHANGELOG.md
@@ -23,3 +23,4 @@
 ## pre-Beta 1.3
 - Recompiled lib.zip
 - Added unload function
+- Recoded globalplaceholders.py